# proc: Caching type offsets