--- a/proc/proc_test.go
+++ b/proc/proc_test.go
@@ -1626,7 +1626,7 @@ func TestPackageVariables(t *testing.T) {
 
 func TestIssue149(t *testing.T) {
 	ver, _ := ParseVersionString(runtime.Version())
-	if ver.Major	> 0 && !ver.AfterOrEqual(GoVersion{1, 7, 0, 0, 0}) {
+	if ver.Major > 0 && !ver.AfterOrEqual(GoVersion{1, 7, 0, 0, 0}) {
 		return
 	}
 	// setting breakpoint on break statement